Working read command
The error I was making was not quite understanding
how to use the datasheet. The address for the various
enable functions exist at 0x80, but require bitlevel
operations to work. I write 1 | 1<<2 to turn on and
enable the device.

What should happen next is a scheme for more granular
state management. For example, I should power-on and enable
the device on a file open.